Updated design

+++ RN_SDK/src/main/java/com/example/rnsdk/Adapter/TransactionHistoryAdapter.java
@@ -83,7 +83,7 @@ public class TransactionHistoryAdapter extends RecyclerView.Adapter<TransactionH
         holder.textBalance.setText("" + history.getBalance());
         holder.textLocation.setText("" + history.getLocationName());
 
-        int point = Integer.parseInt(history.getPoints());
+        Double point = Double.parseDouble(history.getPoints());
         if (point > 0) {
             holder.textPoint.setText("+" + history.getPoints());
         } else {
